project_management/requirements.txt
asgiref==3.11.1
Django==6.0.2
djangorestframework==3.16.1
openpyxl==3.1.5
psycopg[binary]==3.2.6
sqlparse==0.5.5
tzdata==2025.3
gunicorn==23.0.0
